refact(*): improve organization of code
Improve code readability and organization by moving
files into packages.
