+++ world_map/world_map_test.go
@@ -1,4 +1,4 @@
-package main
+package worldmap
 
 import (
 	"testing"
